Removed invalid cast.

+++ src/main/java/org/orekit/propagation/numerical/cr3bp/CR3BPMultipleShooter.java
@@ -143,7 +143,7 @@ public class CR3BPMultipleShooter extends AbstractMultipleShooting {
         }
 
         for (int index : mapConstraints.keySet()) {
-            final int np = (int) (index / 6);
+            final int np = index / 6;
             final int nc = index % 6;
             final AbsolutePVCoordinates absPv = patchedSpacecraftStates.get(np).getAbsPVA();
             if (nc < 3) {
